refactor: implement professional project structure and naming conventions
 Project Structure Improvements:
- Add __tests__/ directory with organized test structure
- Create deploy/ folder for deployment configurations
- Reorganize assets into images/ and icons/logos/ subdirectories
- Rename lib/ to utils/ following community standards
- Consolidate App.css and index.css into globals.css

 Naming Convention Standardization:
- Convert UI components to PascalCase (Button.jsx, Card.jsx, Badge.jsx)
- Rename JSON files to kebab-case (case-studies.json, work-process.json, tech-stack.json)
- Update all imports to reflect new naming conventions

 Documentation Updates:
- Update README.md with new project structure
- Add Project Organization section with naming conventions
- Document file structure standards and code quality practices
- Add test guidelines in __tests__/README.md

 Technical Improvements:
- Update all component imports for new file structure
- Maintain Vercel deployment functionality with config copy
- Ensure consistent code organization across the project
- Prepare foundation for future testing implementation

This refactoring establishes professional standards and improves maintainability while preserving all existing functionality.
