+++ __tests__/README.md
@@ -1,0 +1,41 @@
+# Tests
+
+This directory contains all test files for the project.
+
+## Structure
+
+```
+__tests__/
+├── components/     # Component tests
+├── utils/         # Utility function tests
+└── README.md      # This file
+```
+
+## Running Tests
+
+```bash
+# Run all tests
+npm test
+
+# Run tests in watch mode
+npm run test:watch
+
+# Run tests with coverage
+npm run test:coverage
+```
+
+## Test Guidelines
+
+- Use descriptive test names
+- Follow the AAA pattern (Arrange, Act, Assert)
+- Mock external dependencies
+- Test edge cases and error scenarios
+- Maintain good test coverage
+
+## Future Setup
+
+When ready to implement tests, consider:
+- Installing Jest and React Testing Library
+- Setting up test utilities
+- Configuring test environment
+- Adding test scripts to package.json
