+++ src/assets/images/powerbi_logo.png
