+++ src/assets/images/python_logo.png
