+++ src/components/sections/CaseStudySection.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Button } from '../ui/button';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
-import { Badge } from '../ui/badge';
+import { Button } from '../ui/Button';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
+import { Badge } from '../ui/Badge';
 import { CheckCircle, ExternalLink, MessageCircle, Target, TrendingUp, Zap, Star, ArrowRight } from 'lucide-react';
 import { useSite } from '../../context/SiteContext';
