+++ src/components/sections/CredibilitySection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '../ui/card';
-import { Badge } from '../ui/badge';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '../ui/Card';
+import { Badge } from '../ui/Badge';
 import { TrendingUp, Shield, Award, Brain } from 'lucide-react';
 import { useSite } from '../../context/SiteContext';
