+++ src/components/sections/PartnershipsSection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
-import { Button } from '../ui/button';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
+import { Button } from '../ui/Button';
 import { CheckCircle } from 'lucide-react';
 import { useSite } from '../../context/SiteContext';
