+++ src/components/sections/ServicesSection.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/card';
-import { Badge } from '../ui/badge';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
+import { Badge } from '../ui/Badge';
 import { CheckCircle } from 'lucide-react';
 import { useSite } from '../../context/SiteContext';
