+++ src/components/sections/WorkProcessSection.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '../ui/card';
-import { Button } from '../ui/button';
-import { Badge } from '../ui/badge';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '../ui/Card';
+import { Button } from '../ui/Button';
+import { Badge } from '../ui/Badge';
 import { CheckCircle, Code, Database, Shield, Target, Users, Clock } from 'lucide-react';
 import { useSite } from '../../context/SiteContext';
